+++ framework/model-test/src/main/java/is/codion/framework/model/test/TestDomain.java
@@ -152,8 +152,8 @@ public final class TestDomain extends DomainModel {
 										Detail.INT_DERIVED.as()
 														.derived()
 														.from(Detail.INT)
-														.with(source -> {
-															Integer intValue = source.get(Detail.INT);
+														.with(values -> {
+															Integer intValue = values.get(Detail.INT);
 															if (intValue == null) {
 																return null;
 															}
@@ -319,19 +319,19 @@ public final class TestDomain extends DomainModel {
 										Derived.INT2.as()
 														.derived()
 														.from(Derived.INT1)
-														.with(source -> source.optional(Derived.INT1)
+														.with(values -> values.optional(Derived.INT1)
 																		.map(value -> value + 1)
 																		.orElse(null)),
 										Derived.INT3.as()
 														.derived()
 														.from(Derived.INT2)
-														.with(source -> source.optional(Derived.INT2)
+														.with(values -> values.optional(Derived.INT2)
 																		.map(value -> value + 1)
 																		.orElse(null)),
 										Derived.INT4.as()
 														.derived()
 														.from(Derived.INT3)
-														.with(source -> source.optional(Derived.INT3)
+														.with(values -> values.optional(Derived.INT3)
 																		.map(value -> value + 1)
 																		.orElse(null)))
 						.build();